revisit ui and add ga
