+++ app/components/ui/input.tsx
@@ -9,17 +9,18 @@ export interface InputProps
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
   ({ className, type, leftIcon, ...props }, ref) => {
     return (
-      <div className="relative flex items-center w-full">
+      <div className="relative w-full">
         {leftIcon && (
-          <div className="absolute left-2 flex items-center justify-center">
+          <div className="absolute left-3 top-1/2 -translate-y-1/2">
             {leftIcon}
           </div>
         )}
         <input
           type={type}
           className={cn(
-            "flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-1 text-sm shadow-sm transition-colors file:border-0 file:bg-transparent file:text-sm file:font-medium file:text-foreground placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50",
-            leftIcon && "pl-9",
+            "flex h-10 w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background file:border-0 file:bg-transparent file:text-sm file:font-medium placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 disabled:cursor-not-allowed disabled:opacity-50",
+            leftIcon && "pl-10",
+            "dark:bg-white dark:text-gray-900",
             className
           )}
           ref={ref}
